Fix stupid error

+++ branches/british-isles/libs/event_handlers/every_day/RebuildCategoryStats.class.php
@@ -54,14 +54,22 @@ class RebuildCategoryStats extends EventHandler
 				ENGINE=MyISAM
 				SELECT CRC32(LOWER(imageclass)) AS category_id,imageclass,count(*) AS c,gridimage_id
 				FROM gridimage_search
-				GROUP BY imageclass"); //the autoincrement column doesnt need 'null' for some reason. 
+				GROUP BY imageclass");
 		
-		$db->Execute("DROP TABLE IF EXISTS category_stat");
-		$db->Execute("RENAME TABLE category_stat_tmp TO category_stat");
+		$data = $db->getRow("SHOW TABLE STATUS LIKE 'category_stat_tmp'");
+		if (!empty($data['Create_time']) && strtotime($data['Create_time']) > (time() - 60*15)) {
+
 		
-		//return true to signal completed processing
-		//return false to have another attempt later
-		return true;
+			$db->Execute("DROP TABLE IF EXISTS category_stat");
+			$db->Execute("RENAME TABLE category_stat_tmp TO category_stat");
+		
+		
+			//return true to signal completed processing
+			//return false to have another attempt later
+			return true;
+		} else {
+			return false;
+		}
 	}
 	
 }
